+++ lib/src/client.g.edited.dart
@@ -14,6 +14,35 @@ class _PteroClient implements PteroClient {
   String? baseUrl;
 
   @override
+  Future<void> login(credentials) async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    _data.addAll(credentials.toJson());
+    await _dio.fetch<void>(_setStreamType<void>(
+        Options(method: 'POST', headers: _headers, extra: _extra)
+            .compose(_dio.options, '/auth/login',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    return;
+  }
+
+  @override
+  Future<void> logout() async {
+    const _extra = <String, dynamic>{};
+    final queryParameters = <String, dynamic>{};
+    final _headers = <String, dynamic>{};
+    final _data = <String, dynamic>{};
+    await _dio.fetch<Map<String, dynamic>>(_setStreamType<void>(
+        Options(method: 'POST', headers: _headers, extra: _extra)
+            .compose(_dio.options, '/auth/logout',
+                queryParameters: queryParameters, data: _data)
+            .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
+    return;
+  }
+
+  @override
   Future<FractalResponseList<Server>> getServers() async {
     const _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
@@ -119,7 +148,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/account/two-factor',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -134,7 +163,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/account/email',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -149,7 +178,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/account/password',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -200,7 +229,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/account/api-keys/$apiKey',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -275,7 +304,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/command',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -290,7 +319,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/power',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -366,7 +395,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/databases/$database',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -435,7 +464,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/files/rename',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -450,7 +479,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/files/copy',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -465,7 +494,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/files/write',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -503,7 +532,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/files/decompress',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -541,7 +570,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/files/create-folder',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -660,7 +689,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/schedules/$schedule',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -720,7 +749,7 @@ class _PteroClient implements PteroClient {
                 '/api/client/servers/$server/schedules/$schedule/tasks/$task',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -918,7 +947,7 @@ class _PteroClient implements PteroClient {
             .compose(_dio.options, '/api/client/servers/$server/users/$user',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -1011,7 +1040,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/backups/$backup',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -1067,7 +1096,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/settings/rename',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   @override
@@ -1082,7 +1111,7 @@ class _PteroClient implements PteroClient {
                 _dio.options, '/api/client/servers/$server/settings/reinstall',
                 queryParameters: queryParameters, data: _data)
             .copyWith(baseUrl: baseUrl ?? _dio.options.baseUrl)));
-    return null;
+    return;
   }
 
   RequestOptions _setStreamType<T>(RequestOptions requestOptions) {